+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Other blocks are:
 
 ### Fixed
 
+- **LIB** - Fixed support of multiple spectra in Gaussian log files.
 - **LIB** - Fixed deprecation warning from `Matplotlib` on X/Y scale.
 - **API** - Anharmonic vibrational energies correctly extracted in absence of intensities (DCPT2, HDCPT2).
 - **API** - Support numbers like `2e6` as floating point numbers.